pyproject.toml
[project]
name = "cerebro"
version = "2.0.0"
description = "Cerebro — 群体智能协同引擎"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
cerebro = "cerebro.app:main"
